+++ test/test_api_parse_1.js
@@ -3,7 +3,6 @@
 const clasp = require('../index');
 
 const assert = require('assert');
-const mstreams = require('memory-streams');
 
 function _isPlainObject(obj) {